requirements.txt
requests==2.31.0
urllib3==2.0.7
pandas==2.1.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
